feat(TP-570): add OCP bad example
We have a component that validates the display of a link or button through a type, but if we want to add a new element to display in that section, we will have to modify the component directly. For which it does not comply with the principle, it cannot be extended.