feat(TP-570): add LSP bad example
We have a component that extends the functionality of another, but does not maintain the same props, so it is not replaceable by the previous one. For this reason, it does not comply with the principle since it must be replaceable without the need to affect functionality or give an error.