feat(570): add ISP bad example
We have a component that receives a post object from props and several of its internal components need some values from that object, but we do not only pass that value to the child component but we pass the entire object. For this reason he does not meet the principle since the child component receives more information than he needs.